message-processing-service/pyproject.toml
[project]
name = "message-processing-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.16",
    "fastapi>=0.115.12",
    "openai>=1.72.0",
    "poethepoet>=0.33.1",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]

[tool.poe.tasks]
start = "python -m main"
